pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Basic SEO site checker app. Allow to show if site provide basic info for search engine"
authors = ["Vladislav <balenkov53@gmail.com>"]
readme = "README.md"
packages = [
	{ include = "page_analyzer"}
]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.1.0"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"
validators = "^0.34.0"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
